pyproject.toml
[tool.poetry]
name = "coreapp"
version = "0.1.0"
description = ""
authors = ["Anton <antonzakhlebayeu@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
djangorestframework = "^3.13.1"
python-dotenv = "^0.20.0"
psycopg2 = "^2.9.3"
PyJWT = "^2.4.0"
reference = "^0.0.3"
drf-nested-routers = "^0.93.4"
celery = "^5.2.7"
rabbitmq = "^0.2.0"
black = "^22.6.0"
flake8 = "^4.0.1"
isort = "^5.10.1"
pytest = "^7.1.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"
model-bakery = "^1.6.0"
apiclient = "^1.0.4"
uvicorn = "^0.18.2"
fastapi = "^0.78.0"
motor = "^3.0.0"
decouple = "^0.0.7"
pymongo = "^4.1.1"
pika = "^1.3.0"
aiormq = "^6.3.4"
aio-pika = "^8.0.3"
coreschema = "^0.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '.pyi?$'
exclude = '''
/(
    .git
    | .hg
    | .mypy_cache
    | .tox
    | .venv
    | _build
    | buck-out
    | build
    | dist
)/
'''
